Add slow speeds (0.25x, 0.5x) and improve state diagram contrast
- Add 0.25x and 0.5x speed options to the simulator
- Fix getSpeed() to use parseFloat instead of parseInt
- Give each process state a distinct fill color when active (blue/amber/green/red/purple)

+++ js/simulator.js
@@ -57,7 +57,7 @@
   }
 
   function getSpeed() {
-    return parseInt(document.getElementById('speed-select').value) || 2;
+    return parseFloat(document.getElementById('speed-select').value) || 2;
   }
 
   function wireQuantumVisibility() {
